+++ boxlet/opengl/renderers/sprite_palette_instanced_renderer.py
