+++ boxlet/opengl/renderers/sprite_renderer.py
